One-way platforms

+++ Assets/Scripts/HurtZone.cs
@@ -5,33 +5,33 @@ using UnityEngine;
 public class HurtZone : MonoBehaviour
 {
 
-    private Player player;
+	private Player player;
 
 	private void Start()
 	{
-        player = transform.parent.GetComponent<Player>();
+		player = transform.parent.GetComponent<Player>();
 	}
 
-    private void OnCollisionEnter2D(Collision2D collision)
-    {
-        GameObject other = collision.gameObject;
+	private void OnCollisionEnter2D(Collision2D collision)
+	{
+		GameObject other = collision.gameObject;
 
-        if (other.layer == LayerMask.NameToLayer("LevelGeometry"))
-        {
-            player.Kill();
-        }
-    }
+		if (other.layer == LayerMask.NameToLayer("LevelGeometry") && other.GetComponent<PlatformEffector2D>() == null)
+		{
+			player.Kill();
+		}
+	}
 
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        GameObject other = collision.gameObject;
+	private void OnTriggerEnter2D(Collider2D collision)
+	{
+		GameObject other = collision.gameObject;
 
-        Damage damage = other.GetComponent<Damage>();
-        if (damage != null)
-        {
-            //TODO: damage, respawn if necessary (like for spikes)
-            player.Kill();
-        }
+		Damage damage = other.GetComponent<Damage>();
+		if (damage != null)
+		{
+			//TODO: damage, respawn if necessary (like for spikes)
+			player.Kill();
+		}
 
-    }
+	}
 }
